Change price data type to float in struct def

+++ header.h
@@ -3,5 +3,5 @@ typedef struct {
     char description[50];   //Item Description
     int qty;                //Item Quantity
     char exp[10];           //Item Expiry Date
-    double price;           //Item price
+    float price;           //Item price
   } item;
